requirements.txt
asgiref==3.5.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
Django==3.2.9
django-cors-headers==3.10.0
mysqlclient==2.1.0
pycparser==2.21
PyJWT==2.3.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
